This adds SR flip-flop
Flip-flop with set and reset input signals. Clock input signal is optional. It can be triggered by raising or failing edge. Master reset (which pulls output low regardless of the state of S and R) is also optional and can be inverted. Such SR flip-flop with inverted master reset is used internally in 555 chip.