Add generic smart relay component
I work at a Tafe in Australia and we are after a simple circuit diagram creator, but many of the components we use are not electronic but do get represented via a circuit diagram. I hope to continue to add to this library so that I can use with classes of students to help them build some basic circuit diagrams. Or at least be exposed to the concept.